pyproject.toml
[tool.poetry]
name = "bintest"
version = "0.2.0"
description = "python library for test binarys"
authors = ["Artur Gomes <contato@arturgomes.com.br>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/arturgoms/python-bintest"
repository = "https://github.com/arturgoms/python-bintest"
keywords = ["unittest", "binary"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
scripttest = "^1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
